Split KeysBackup to several files. No other change.
